✨ Extract dialog to <App> so can be invoked by every page
