+++ src/pages/AgendaPage.vue
@@ -18,12 +18,6 @@
         agenda-key-of-start-time="START"
         agenda-key-of-end-time="END"
       />
-
-      <AgendumDialog
-        :open="!!dialogAgendum"
-        :agendum="dialogAgendum"
-        @close="closeDialog"
-      />
     </section>
 
     <section id="maps" class="cell large-10">
@@ -40,13 +34,10 @@
 </template>
 
 <script>
-import router from '@/router'
 import CapsuleRadioButton from '@/components/CapsuleRadioButton'
 import ParallelAgenda from '@/components/ParallelAgenda'
 import AgendumCell from '@/components/AgendumCell'
-import AgendumDialog from '@/components/AgendumDialog'
 import get from 'lodash/get'
-import find from 'lodash/find'
 import filter from 'lodash/filter'
 import sortBy from 'lodash/sortBy'
 import groupBy from 'lodash/groupBy'
@@ -58,7 +49,6 @@ export default {
     CapsuleRadioButton,
     ParallelAgenda,
     AgendumCell,
-    AgendumDialog,
   },
   data() {
     let schedule = POPULATED_SCHEDULE.map(record => ({
@@ -77,26 +67,11 @@ export default {
       dates,
       activeDate: dates[0],
       AgendumCell,
-      dialogAgendum: find(POPULATED_SCHEDULE, [
-        'id',
-        this.$route.params.agendumId,
-      ]),
     }
-  },
-  watch: {
-    $route(to, from) {
-      const { agendumId } = to.params
-      this.dialogAgendum = agendumId
-        ? find(POPULATED_SCHEDULE, ['id', agendumId])
-        : null
-    },
   },
   methods: {
     agendumPropsMapper(agendum) {
       return agendum
-    },
-    closeDialog() {
-      router.push({ name: 'AgendaPage' })
     },
     formatDate(date) {
       return date.toLocaleDateString('en-US', {
